packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared domain models, repositories, and infrastructure"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.9.0",
    "pydantic-settings>=2.4.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
    "redis>=5.2.0",
    "structlog>=24.4.0",
    "boto3>=1.35.0",
    "cryptography>=44.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "tenacity>=8.2.3",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"